show viewers, amount of likes

+++ backend/src/viewers/mod.rs
@@ -6,7 +6,7 @@ use async_trait::async_trait;
 
 #[async_trait]
 pub trait Viewers: Send + Sync {
-    async fn count(&self, key: &str) -> isize;
+    async fn count(&self, key: &str) -> i64;
     async fn add(&self, key: &str);
     async fn remove(&self, key: &str);
 }
